gal-451# made some changes for validation and added more unit test
